+ update Methode aller Controller überarbeitet + delete Methode aller Controller/Services überarbeitet

+++ src/main/java/de/questor/services/QuestMarkerService.java
@@ -29,7 +29,11 @@ public class QuestMarkerService {
         return questMarkerRepository.save(questMarker);
     }
 
-    public void delete(Integer id) {
-        questMarkerRepository.deleteById(id);
+    public boolean delete(Integer id) {
+        if (questMarkerRepository.existsById(id)) {
+            questMarkerRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
